// src/types/ast.ts
export enum BlendMode {
	// Normal blends:
	/** (Only applicable to objects with children) */
	PASS_THROUGH = "PASS_THROUGH",
	/** (Only applicable to objects with children) */
	NORMAL = "NORMAL",

	/** Darken */
	DARKEN = "DARKEN",
	MULTIPLY = "MULTIPLY",
	LINEAR_BURN = "LINEAR_BURN",
	COLOR_BURN = "COLOR_BURN",

	/** Lighten */
	LIGHTEN = "LIGHTEN",
	SCREEN = "SCREEN",
	LINEAR_DODGE = "LINEAR_DODGE",
	COLOR_DODGE = "COLOR_DODGE",

	/** Contrast */
	OVERLAY = "OVERLAY",
	SOFT_LIGHT = "SOFT_LIGHT",
	HARD_LIGHT = "HARD_LIGHT",

	/** Inversion */
	DIFFERENCE = "DIFFERENCE",
	EXCLUSION = "EXCLUSION",

	/** Component */
	HUE = "HUE",
	SATURATION = "SATURATION",
	COLOR = "COLOR",
	LUMINOSITY = "LUMINOSITY",
}
// ...
export enum EffectType {
	INNER_SHADOW = "INNER_SHADOW",
	DROP_SHADOW = "DROP_SHADOW",
	LAYER_BLUR = "LAYER_BLUR",
	BACKGROUND_BLUR = "BACKGROUND_BLUR",
}
// ...
export interface DOCUMENT {
	/** An array of canvases attached to the document */
	children: Node[];
}

/** Represents a single page */
export interface CANVAS {
	/** An array of top level layers on the canvas */
	children: Node[];
	/** Background color of the canvas */
	backgroundColor: Color;
	/** default: [] An array of export settings representing images to export from the canvas */
	exportSettings: ExportSetting[];
	/** Node ID that corresponds to the start frame for prototypes */
	prototypeStartNodeID?: string | null;
}

/** A node of fixed size containing other nodes */
export interface FRAME {
	/** An array of nodes that are direct children of this node */
	children: Node[];
	/** If true, layer is locked and cannot be edited, default `false` */
	locked?: boolean;
	/** Background of the node */
	background: Paint[];
	/** Background color of the node. This is deprecated, as frames now support more than a solid color as a background. Please use the background field instead. */
	backgroundColor?: Color;
	/** default: [] An array of export settings representing images to export from node */
	exportSettings: ExportSetting[];
	/** How this node blends with nodes behind it in the scene (see blend mode section for more details) */
	blendMode: BlendMode;
	/** default: false Keep height and width constrained to same ratio */
	preserveRatio: boolean;
	/** Horizontal and vertical layout constraints for node */
	constraints: LayoutConstraint;
	/** default: null Node ID of node to transition to in prototyping */
	transitionNodeID?: string | null;
	/** default: null The duration of the prototyping transition on this node (in milliseconds). */
	transitionDuration?: number | null;
	/** default: null The easing curve used in the prototyping transition on this node. */
	transitionEasing?: EasingType | null;
	/** default: 1 Opacity of the node */
	opacity: number;
	/** Bounding box of the node in absolute space coordinates */
	absoluteBoundingBox: Rectangle;
	/** Width and height of element. This is different from the width and height of the bounding box in that the absolute bounding box represents the element after scaling and rotation. Only present if geometry=paths is passed */
	size?: Vector;
	/** The top two rows of a matrix that represents the 2D transform of this node relative to its parent. The bottom row of the matrix is implicitly always (0, 0, 1). Use to transform coordinates in geometry. Only present if geometry=paths is passed */
	relativeTransform?: Transform;
	/** Does this node clip content outside of its bounds? */
	clipsContent: boolean;
	/** default: [] An array of layout grids attached to this node (see layout grids section for more details). GROUP nodes do not have this attribute */
	layoutGrids?: LayoutGrid[];
	/** default: [] An array of effects attached to this node (see effects section for more details) */
	effects: Effect[];
	/** default: false Does this node mask sibling nodes in front of it? */
	isMask: boolean;
	/** default: false Does this mask ignore fill style (like gradients) and effects? */
	isMaskOutline: boolean;
}

/** A logical grouping of nodes */
export type GROUP = FRAME;
//  {
//     /** How this node blends with nodes behind it in the scene (see blend mode section for more details) */
//     blendMode: BlendMode;
//     children: Node[],
// }

/** A vector network, consisting of vertices and edges */
export interface VECTOR {
	/** default: [] An array of export settings representing images to export from node */
	exportSettings: ExportSetting[];
	/** If true, layer is locked and cannot be edited, default `false` */
	locked?: boolean;
	/** How this node blends with nodes behind it in the scene (see blend mode section for more details) */
	blendMode: BlendMode;
	/** default: false Keep height and width constrained to same ratio */
	preserveRatio?: boolean;
	/** Horizontal and vertical layout constraints for node */
	constraints: LayoutConstraint;
	/** default: null Node ID of node to transition to in prototyping */
	transitionNodeID?: string | null;
	/** default: null The duration of the prototyping transition on this node (in milliseconds). */
	transitionDuration?: number | null;
	/** default: null The easing curve used in the prototyping transition on this node. */
	transitionEasing?: EasingType | null;
	/** default: 1 Opacity of the node */
	opacity?: number;
	/** Bounding box of the node in absolute space coordinates */
	absoluteBoundingBox: Rectangle;
	/** Width and height of element. This is different from the width and height of the bounding box in that the absolute bounding box represents the element after scaling and rotation. Only present if geometry=paths is passed */
	size?: Vector;
	/** The top two rows of a matrix that represents the 2D transform of this node relative to its parent. The bottom row of the matrix is implicitly always (0, 0, 1). Use to transform coordinates in geometry. Only present if geometry=paths is passed */
	relativeTransform?: Transform;
	/** default: [] An array of effects attached to this node (see effects section for more details) */
	effects?: Effect[];
	/** default: false Does this node mask sibling nodes in front of it? */
	isMask?: boolean;
	/** default: [] An array of fill paints applied to the node */
	fills: Paint[];
	/** Only specified if parameter geometry=paths is used. An array of paths representing the object fill */
	fillGeometry?: Path[];
	/** default: [] An array of stroke paints applied to the node */
	strokes: Paint[];
	/** The weight of strokes on the node */
	strokeWeight: number;
	strokeCap?: StrokeCap;
	/** Only specified if parameter geometry=paths is used. An array of paths representing the object stroke */
	strokeGeometry?: Path[];
	/** Where stroke is drawn relative to the vector outline as a string enum
    "INSIDE": draw stroke inside the shape boundary
    "OUTSIDE": draw stroke outside the shape boundary
    "CENTER": draw stroke centered along the shape boundary */
	strokeAlign: StrokeAlign;
	/** A string enum with value of "MITER", "BEVEL", or "ROUND", describing how corners in vector paths are rendered. */
	strokeJoin?: StrokeJoin;
	/** An array of floating point numbers describing the pattern of dash length and gap lengths that the vector path follows. For example a value of [1, 2] indicates that the path has a dash of length 1 followed by a gap of length 2, repeated. */
	strokeDashes?: number[];
	/** Only valid if strokeJoin is "MITER". The corner angle, in degrees, below which strokeJoin will be set to "BEVEL" to avoid super sharp corners. By default this is 28.96 degrees. */
	strokeMiterAngle?: number;
	/** A mapping of a StyleType to style ID (see Style) of styles present on this node. The style ID can be used to look up more information about the style in the top-level styles field. */
	styles?: StylesMap;
}

/** A group that has a boolean operation applied to it */
export type BOOLEAN = VECTOR & {
	/** An array of nodes that are being boolean operated on */
	children: Node[];
};

/** A group that has a boolean operation applied to it */
export type BOOLEAN_OPERATION = VECTOR & {
	/** An array of nodes that are being boolean operated on */
	children: Node[];
	/** A string enum with value of "UNION", "INTERSECT", "SUBTRACT", or "EXCLUDE" indicating the type of boolean operation applied */
	booleanOperation: BooleanOperationType;
};

/** A regular star shape */
export type STAR = VECTOR;

/** A straight line */
export type LINE = VECTOR;

/** An ellipse */
export type ELLIPSE = VECTOR;

/** A regular n-sided polygon */
export type REGULAR_POLYGON = VECTOR;

/** A rectangle */
export type RECTANGLE = VECTOR & {
	/** Radius of each corner of the rectangle */
	cornerRadius: number;
	/** Array of length 4 of the radius of each corner of the rectangle, starting in the top left and proceeding clockwise */
	rectangleCornerRadii: [number, number, number, number];
};

/** A text box */
export type TEXT = VECTOR & {
	/** Text contained within text box */
	characters: string;
	/** Style of text including font family and weight (see type style section for more information) */
	style: TypeStyle;
	/** Array with same number of elements as characeters in text box, each element is a reference to the styleOverrideTable defined below and maps to the corresponding character in the characters field. Elements with value 0 have the default type style */
	characterStyleOverrides: number[];
	/** Map from ID to TypeStyle for looking up style overrides */
	styleOverrideTable: { [mapId: number]: TypeStyle };
};

/** A rectangular region of the canvas that can be exported */
export interface SLICE {
	/** An array of export settings representing images to export from this node */
	exportSettings: ExportSetting[];
	/** Bounding box of the node in absolute space coordinates */
	absoluteBoundingBox: Rectangle;
	/** Width and height of element. This is different from the width and height of the bounding box in that the absolute bounding box represents the element after scaling and rotation. Only present if geometry=paths is passed */
	size?: Vector;
	/** The top two rows of a matrix that represents the 2D transform of this node relative to its parent. The bottom row of the matrix is implicitly always (0, 0, 1). Use to transform coordinates in geometry. Only present if geometry=paths is passed */
	relativeTransform?: Transform;
}

/** A node that can have instances created of it that share the same properties */
export type COMPONENT = FRAME;

/** A node that can have instances created of it that share the same properties */
export type COMPONENT_SET = FRAME;

/** An instance of a component, changes to the component result in the same changes applied to the instance */
export type INSTANCE<ComponentID = string> = FRAME & {
	/** ID of component that this instance came from, refers to components table (see endpoints section below) */
	componentId: ComponentID;
};

export type NodeTypes = {
	/** The root node */
	DOCUMENT: DOCUMENT;
	/** Represents a single page */
	CANVAS: CANVAS;
	/** A node of fixed size containing other nodes */
	FRAME: FRAME;
	/** A logical grouping of nodes */
	GROUP: GROUP;
	/** A vector network, consisting of vertices and edges */
	VECTOR: VECTOR;
	/** A group that has a boolean operation applied to it */
	BOOLEAN: BOOLEAN;
	BOOLEAN_OPERATION: BOOLEAN_OPERATION;
	/** A regular star shape */
	STAR: STAR;
	/** A straight line */
	LINE: LINE;
	/** An ellipse */
	ELLIPSE: ELLIPSE;
	/** A regular n-sided polygon */
	REGULAR_POLYGON: REGULAR_POLYGON;
	/** A rectangle */
	RECTANGLE: RECTANGLE;
	/** A text box */
	TEXT: TEXT;
	/** A rectangular region of the canvas that can be exported */
	SLICE: SLICE;
	/** A node that can have instances created of it that share the same properties */
	COMPONENT: COMPONENT;
	/** A node that can have instances created of it that share the same properties */
	COMPONENT_SET: COMPONENT_SET;
	/** An instance of a component, changes to the component result in the same changes applied to the instance */
	INSTANCE: INSTANCE;
};

export type NodeType = keyof NodeTypes;

export type Node<NType extends NodeType = NodeType> = {
	id: string;
	name: string;
	visible: boolean;
	type: NType;
} & NodeTypes[NType];